Removed the reinstallation of stuff, and we don't need to reboot
The re-installation of the kernel headers doesn't seem necessary,
and hasn't happened the last two times I went through this process
because I wasn't connected to the internet.

I figured out (finally!) that I don't have to reboot the system
to use the new driver, but I only have to modprobe it. Huzzah!!
Signed-off-by: Elliot V <elliotfriend@gmail.com>